add support for `#[repr(packed(n))]`

+++ tests/test_struct.rs
@@ -10,9 +10,9 @@ pub struct Child1 {
 }
 
 #[derive(VolatileStruct)]
-#[repr(C)]
+#[repr(C, packed)]
 pub struct Child2 {
-    field1: u32,
+    field1: u8,
     field2: u32,
     field3: u32,
 }
@@ -39,7 +39,7 @@ pub fn sum_parent(parent: &VolatileParent) -> u32 {
 }
 
 pub fn sum_child2(child: &VolatileChild2) -> u32 {
-    child.field1.read() + child.field2.read() + child.field3.read()
+    u32::from(child.field1.read()) + child.field2.read() + child.field3.read()
 }
 
 pub fn modify(parent: &mut VolatileParent) {
